fixed float printing

+++ cimport.c
@@ -57,10 +57,10 @@ int putchar_helper(struct Function_* closure, int c) {
 }
 
 float int_to_float_helper(struct Function_* closure, int toCast) {
-  return ((double) toCast);
+  return ((float) toCast);
 }
 
-int float_to_int_helper(struct Function_* closure, double toCast) {
+int float_to_int_helper(struct Function_* closure, float toCast) {
   return ((int) toCast);
 }
